+++ src/arpes/analysis/tarpes.py
@@ -10,7 +10,7 @@ import xarray as xr
 
 from arpes.debug import setup_logger
 from arpes.preparation import normalize_dim
-from arpes.preparation.axis_preparation import vstack_data
+from arpes.preparation.axis import vstack_data
 from arpes.provenance import update_provenance
 from arpes.utilities import normalize_to_spectrum